pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp_api_auth"
version = "0.1.0"
description = "Sample MCP server with API key authentication and demo tools"
authors = [
  { name = "Cascade AI" }
]
dependencies = [
  "mcp>=0.1.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "starlette>=0.37.0",
  "sse-starlette>=2.0.0",
  "httpx>=0.27.0",
  "uvicorn>=0.30.0",
  "fastmcp>=2.12.4",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
  "ruff>=0.5.5"
]

[project.scripts]
mcp-api-auth-server = "mcp_api_auth.server:main"
